api/requirements.txt
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.5.0
matplotlib
numpy